kernel-ark: set kernelconfigs via kernel-ark-config repo
This patch drops the current way of setting kernel configfs for
kernel-ark project. It is confirmed by checking the rpm kernelconfig
file that those configs are not activated.

The hopefully right way is to follow instruction [0] that describes to
put additional kernelconfigs into custom-overrides/generic or
custom-overrides/debug. Those directories are always empty and made for
possible user changes and should be "mergable" without any conflicts to
activate additional configs that will be considered by the
build_configs.sh script.

Also we activate DEBUG_ATOMIC_SLEEP to notify over the kernel log if
there are any cases when we sleep while atomic context.

In combination with the sosreports it is easy to find any KASAN, UBSAN
or atomic sleep issues.

[0] https://gitlab.com/cki-project/kernel-ark/-/tree/os-build/redhat/configs