kernel-ark: use merge for ci_test_config branch
This patch introduce a for each loop to merge all gfs2, dlm and
kernel-ark config branches. This should fix the issue with rebase
ci_test_config branch on kernel-ark/os-build branch since it got a
forces push. Now ci_test_config is an upstream Linux kernel repository
(which should not have any changes in history with os-build) that
contains the redhat/configs/custom-overrides changes which should never
fail to run into a conflict with os-build branch.