pacemaker: update configure and rpm options
RPMDEST has defaulted to subtree since 2.1.0, so we don't need that anymore.

--with-cibsecrets can be on any distro, not just rhel8, though it's a good idea
to skip one distro just to ensure the build works without it.

--enable-nls is a newer option (available in all branches) worth testing, again
skipping one distro.

--with-concurrent-fencing-default=true should be done only for rhel with the
2.1 branch; true will become the "default default" in main and 3.0 soon, and
changing it now doesn't hurt anything.

rhel8 with the 2.1 branch should get --enable-compat-2.0 and
--enable-legacy-links for pcs compatibility; other distros/branches don't need
it (the option is ignored in main and 3.0).

Lastly, it's worth building RPMs --with linuxha on SUSE so that code at least
gets compiled sometime.