Fix failure while building pyparsing on RHEL 9
On RHEL 9, there have been consistent failures in CI because installation of pyparsing done by autotools fails to build the pyparsing wheel. The issue is in the license field of pyproject.toml which has been changed and is incompatible with some releases of flit-core which is the build backend of pyparsing. The version of flit-core installed from RHEL 9 repositories has an issue with the new license field.